描画先によらず共通となるレンダリング処理をRendererに分離
